+++ include/wmath.h
@@ -10,19 +10,19 @@
 #include <stdint.h>
 #include <math.h>
 
-void linspace(double start, double end, double phase_angle, double *array);
+void linspace(float start, float end, float phase_angle, float *array);
 
-int squeeze_array(double *in, double *out, size_t N_in, size_t N_out);
+int squeeze_array(float *in, float *out, size_t N_in, size_t N_out);
 
-void norm2_v(kiss_fft_cpx *c_in, double *a_out, uint32_t N);
+void norm2_v(kiss_fft_cpx *c_in, float *a_out, uint32_t N);
 
-void init_gaussian(int n_samples, double* array, double amplitude, double variance);
+void init_gaussian(int n_samples, float* array, float amplitude, float variance);
 
-void apply_gaussian(int n_samples, double *array, double *kernel);
+void apply_gaussian(int n_samples, float *array, float *kernel);
 
-void shift_column_left(double m[N_VERT_LINES][N_SAMPLES_OUT]);
+void shift_column_left(float m[N_VERT_LINES][N_SAMPLES_OUT]);
 
-static inline double norm2(kiss_fft_cpx* c) {
+static inline float norm2(kiss_fft_cpx* c) {
   return sqrt(c->r * c->r + c->i * c->i);
 }
